Clean up user-algebra impl
