+++ algebra-user/src/main/scala/pms/algebra/user/UserAlgebra.scala
@@ -17,5 +17,5 @@ trait UserAlgebra[F[_]] {
 object UserAlgebra {
   import pms.effects._
 
-  def async[F[_]: Async](implicit transactor: Transactor[F]): UserAlgebra[F] = new impl.AsyncAlgebraImpl[F]()
+  def async[F[_]: Async](implicit transactor: Transactor[F]): UserAlgebra[F] = new impl.UserAlgebraImpl[F]()
 }
