+++ algebra-user/src/main/scala/pms/algebra/user/UserAuthAlgebra.scala
@@ -11,7 +11,7 @@ import doobie.util.transactor.Transactor
   * @since 20 Jun 2018
   *
   */
-abstract class UserAuthAlgebra[F[_]](implicit val monadError: MonadError[F, Throwable]) {
+abstract class UserAuthAlgebra[F[_]](implicit private val monadError: MonadError[F, Throwable]) {
 
   def authenticate(email: Email, pw: PlainTextPassword): F[AuthCtx]
 
@@ -73,5 +73,5 @@ abstract class UserAuthAlgebra[F[_]](implicit val monadError: MonadError[F, Thro
 }
 
 object UserAuthAlgebra {
-  def async[F[_]: Async](implicit transactor: Transactor[F]): UserAuthAlgebra[F] = new impl.AsyncAlgebraImpl[F]()
+  def async[F[_]: Async](implicit transactor: Transactor[F]): UserAuthAlgebra[F] = new impl.UserAlgebraImpl[F]()
 }
